# static/swagger.yaml
---
openapi: 3.0.2
info:
  title: YamlWareHouse
  version: 0.1.0
paths:
  "/validate":
    post:
      summary: Validate Yaml document correctness
      operationId: validate_validate_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  "/yaml":
    get:
      summary: Get yaml docs List
      operationId: read_yaml_list_yaml_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      summary: Create Yaml doc
      operationId: create_yaml_yaml_post
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
  "/yaml/{yaml_doc_id}":
    get:
      summary: Get yaml doc by Id
      operationId: read_yaml_yaml__yaml_doc_id__get
      parameters:
      - required: true
        schema:
          title: Yaml Document Id
          type: string
        name: yaml_doc_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    put:
      summary: Update Yaml
      operationId: update_yaml_yaml__yaml_doc_id__put
      parameters:
      - required: true
        schema:
          title: Yaml Doc Id
          type: string
        name: yaml_doc_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Yaml
      operationId: delete_yaml_yaml__yaml_doc_id__delete
      parameters:
      - required: true
        schema:
          title: Yaml Doc Id
          type: string
        name: yaml_doc_id
        in: path
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
components:
  schemas:
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            "$ref": "#/components/schemas/ValidationError"
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
            - type: string
            - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string